+++ src/modules/adsb-feeder/filesystem/root/opt/adsb/adsb-setup/utils/other_aggregators.py
@@ -83,7 +83,9 @@ class Aggregator:
                     capture_output=True,
                 )
             except subprocess.TimeoutExpired as exc2:
-                print_err(f"failed to remove the container {name} stderr: {str(exc2.stdout)} / stdout: {str(exc2.stderr)}")
+                print_err(
+                    f"failed to remove the container {name} stderr: {str(exc2.stdout)} / stdout: {str(exc2.stderr)}"
+                )
 
         # let's make sure the container isn't still there, if it is the docker run won't work
         force_remove_container("temp_container")
@@ -324,7 +326,7 @@ class FlightAware(Aggregator):
 class RadarBox(Aggregator):
     def __init__(self, system: System):
         super().__init__(
-            name="RadarBox",
+            name="AirNav Radar",
             tags=["radarbox"],
             system=system,
         )